Revert previous change and include separate validation functions for helmrepo deployment and the mcm deployments. This adds extra duplication, but is important if one of them, or any new deployment, requires custom validation logic.
