+++ pkg/helmrepo/helmrepo.go
@@ -12,6 +12,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 // Name of helm repo
@@ -137,4 +138,57 @@ func Service(m *operatorsv1alpha1.MultiClusterHub) *corev1.Service {
 		*metav1.NewControllerRef(m, m.GetObjectKind().GroupVersionKind()),
 	})
 	return s
+}
+
+// ValidateDeployment returns a deep copy of the deployment with the desired spec based on the MultiClusterHub spec.
+// Returns true if an update is needed to reconcile differences with the current spec.
+func ValidateDeployment(m *operatorsv1alpha1.MultiClusterHub, dep *appsv1.Deployment) (*appsv1.Deployment, bool) {
+	var log = logf.Log.WithValues("Deployment.Namespace", dep.GetNamespace(), "Deployment.Name", dep.GetName())
+	found := dep.DeepCopy()
+
+	pod := &found.Spec.Template.Spec
+	container := &found.Spec.Template.Spec.Containers[0]
+	needsUpdate := false
+
+	// verify image pull secret
+	if m.Spec.ImagePullSecret != "" {
+		ps := corev1.LocalObjectReference{Name: m.Spec.ImagePullSecret}
+		if !utils.ContainsPullSecret(pod.ImagePullSecrets, ps) {
+			log.Info("Enforcing imagePullSecret from CR spec")
+			pod.ImagePullSecrets = append(pod.ImagePullSecrets, ps)
+			needsUpdate = true
+		}
+	}
+
+	// verify image repository and suffix
+	if container.Image != Image(m) {
+		log.Info("Enforcing image repo and suffix from CR spec")
+		container.Image = Image(m)
+		needsUpdate = true
+	}
+
+	// verify image pull policy
+	if container.ImagePullPolicy != m.Spec.ImagePullPolicy {
+		log.Info("Enforcing imagePullPolicy from CR spec")
+		container.ImagePullPolicy = m.Spec.ImagePullPolicy
+		needsUpdate = true
+	}
+
+	// verify node selectors
+	desiredSelectors := utils.NodeSelectors(m)
+	if !utils.ContainsMap(pod.NodeSelector, desiredSelectors) {
+		log.Info("Enforcing node selectors from CR spec")
+		pod.NodeSelector = desiredSelectors
+		needsUpdate = true
+	}
+
+	// verify replica count
+	if *found.Spec.Replicas != int32(m.Spec.ReplicaCount) {
+		log.Info("Enforcing replicaCount from CR spec")
+		replicas := int32(m.Spec.ReplicaCount)
+		found.Spec.Replicas = &replicas
+		needsUpdate = true
+	}
+
+	return found, needsUpdate
 }
