+++ pkg/mcm/mcm_test.go
@@ -1,4 +1,4 @@
-package helmrepo
+package mcm
 
 import (
 	"reflect"
@@ -9,60 +9,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-func TestDeployment(t *testing.T) {
-	empty := &operatorsv1alpha1.MultiClusterHub{
-		ObjectMeta: metav1.ObjectMeta{Namespace: "test"},
-		Spec: operatorsv1alpha1.MultiClusterHubSpec{
-			Version:         "",
-			ImageRepository: "",
-			ImagePullPolicy: "",
-			ImagePullSecret: "",
-			ImageTagSuffix:  "",
-			NodeSelector: &operatorsv1alpha1.NodeSelector{
-				OS:                  "",
-				CustomLabelSelector: "",
-				CustomLabelValue:    "",
-			},
-			Mongo: operatorsv1alpha1.Mongo{},
-		},
-	}
-	t.Run("MCH with empty fields", func(t *testing.T) {
-		_ = Deployment(empty)
-	})
-
-	essentialsOnly := &operatorsv1alpha1.MultiClusterHub{
-		ObjectMeta: metav1.ObjectMeta{Namespace: "test"},
-		Spec: operatorsv1alpha1.MultiClusterHubSpec{
-			Version:         "test",
-			ImageRepository: "test",
-			ImagePullPolicy: "test",
-			ImageTagSuffix:  "test",
-		},
-	}
-	t.Run("MCH with only required values", func(t *testing.T) {
-		_ = Deployment(essentialsOnly)
-	})
-}
-
-func TestService(t *testing.T) {
-	mch := &operatorsv1alpha1.MultiClusterHub{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "testName",
-			Namespace: "testNS",
-		},
-	}
-
-	t.Run("Create service", func(t *testing.T) {
-		s := Service(mch)
-		if ns := s.Namespace; ns != "testNS" {
-			t.Errorf("expected namespace %s, got %s", "testNS", ns)
-		}
-		if ref := s.GetOwnerReferences(); ref[0].Name != "testName" {
-			t.Errorf("expected ownerReference %s, got %s", "testName", ref[0].Name)
-		}
-	})
-}
 
 func TestValidateDeployment(t *testing.T) {
 	mch := &operatorsv1alpha1.MultiClusterHub{
@@ -83,7 +29,7 @@ func TestValidateDeployment(t *testing.T) {
 	}
 
 	// 1. Valid mch
-	dep := Deployment(mch)
+	dep := ControllerDeployment(mch)
 
 	// 2. Modified ImagePullSecret
 	dep1 := dep.DeepCopy()
