+++ pkg/utils/utils.go
@@ -5,11 +5,9 @@ import (
 	"time"
 
 	operatorsv1alpha1 "github.com/open-cluster-management/multicloudhub-operator/pkg/apis/operators/v1alpha1"
-	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 const (
@@ -130,57 +128,4 @@ func MchIsValid(m *operatorsv1alpha1.MultiClusterHub) bool {
 		m.Spec.Mongo.ReplicaCount <= 0
 
 	return !invalid
-}
-
-// ValidateDeployment returns a deep copy of the deployment with the desired spec based on the MultiClusterHub spec.
-// Returns true if an update is needed to reconcile differences with the current spec.
-func ValidateDeployment(m *operatorsv1alpha1.MultiClusterHub, dep *appsv1.Deployment, image string) (*appsv1.Deployment, bool) {
-	var log = logf.Log.WithValues("Deployment.Namespace", dep.GetNamespace(), "Deployment.Name", dep.GetName())
-	found := dep.DeepCopy()
-
-	pod := &found.Spec.Template.Spec
-	container := &found.Spec.Template.Spec.Containers[0]
-	needsUpdate := false
-
-	// verify image pull secret
-	if m.Spec.ImagePullSecret != "" {
-		ps := corev1.LocalObjectReference{Name: m.Spec.ImagePullSecret}
-		if !ContainsPullSecret(pod.ImagePullSecrets, ps) {
-			log.Info("Enforcing imagePullSecret from CR spec")
-			pod.ImagePullSecrets = append(pod.ImagePullSecrets, ps)
-			needsUpdate = true
-		}
-	}
-
-	// verify image repository and suffix
-	if container.Image != image {
-		log.Info("Enforcing image repo and suffix from CR spec")
-		container.Image = image
-		needsUpdate = true
-	}
-
-	// verify image pull policy
-	if container.ImagePullPolicy != m.Spec.ImagePullPolicy {
-		log.Info("Enforcing imagePullPolicy from CR spec")
-		container.ImagePullPolicy = m.Spec.ImagePullPolicy
-		needsUpdate = true
-	}
-
-	// verify node selectors
-	desiredSelectors := NodeSelectors(m)
-	if !ContainsMap(pod.NodeSelector, desiredSelectors) {
-		log.Info("Enforcing node selectors from CR spec")
-		pod.NodeSelector = desiredSelectors
-		needsUpdate = true
-	}
-
-	// verify replica count
-	if *found.Spec.Replicas != int32(m.Spec.ReplicaCount) {
-		log.Info("Enforcing replicaCount from CR spec")
-		replicas := int32(m.Spec.ReplicaCount)
-		found.Spec.Replicas = &replicas
-		needsUpdate = true
-	}
-
-	return found, needsUpdate
 }
